style fixes and break out studies tests into separate files
